update to vs2019
compiles with vs2019, all warnings fixed, but can't link to BWTA anymore

+++ BOSS/source/BuildOrderSearchGoal.cpp
@@ -140,7 +140,7 @@ bool BuildOrderSearchGoal::isAchievedBy(const GameState & state)
             }
         }
 
-        if (have < getGoal(actionType))
+        if (have < (int)getGoal(actionType))
         {
             return false;
         }
